pyproject.toml
[tool.poetry]

name = "takeshiscastle"
version = "0.0.1"
description = ""
authors = ["Armando Bellante, William Bonvini, Stefano Vanerio"]
readme = "README.md"

[tool.poetry.dependencies]
numpy = "^1.20"
pandas = "^1.4"
scikit-learn = "^1.0"
pillow = "^9.4"
ipython = "^8.11.0"


python = ">=3.9.10, <3.12"
parquet = "^1.3.1"
matplotlib = "^3.7.1"
tqdm = "^4.65.0"
optuna = "^3.1.0"
parallelbar = "^1.3.1"
numba = "^0.56.4"
jinja2 = "^3.1.2"
jupyter-resource-usage = "^0.7.2"
notebook = "^6.5.4"
jupyter-contrib-nbextensions = "^0.7.0"
h5py = "^3.8.0"
fastparquet = "^2023.2.0"
pyarrow = "^11.0.0"



[tool.poetry.group.dev.dependencies]
google-api-python-client = "^2.85.0"
google-auth-httplib2 = "^0.1.0"
google-auth-oauthlib = "^1.0.0"
tabulate = "^0.9.0"
requests = "^2.28.2"
tqdm = "^4.65.0"

[build-system]

requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"



